requirement.txt
asgiref==3.7.2
Django==4.2.6
Faker==19.11.0
Pillow==10.1.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
